Move test helpers to separate files
