+++ EssentialFeed/EssentialFeediOSTests/FeedViewControllerTests.swift
@@ -277,179 +277,4 @@ final class FeedViewControllerTests: XCTestCase {
             assert(sut, hasConfiguredView: image, at: index, file: file, line: line)
         }
     }
-
-    // MARK: - Loader SPY
-
-    final class LoaderSpy: FeedLoader, FeedImageDataLoader {
-
-        // MARK: - FeedLoader
-
-        var loadFeedCallCount: Int {
-            feedRequests.count
-        }
-
-        private var feedRequests = [(FeedLoader.Result) -> Void]()
-
-        func load(completion: @escaping (FeedLoader.Result) -> Void) {
-            feedRequests.append(completion)
-        }
-
-        func completeFeedLoading(with feed: [FeedImage] = [], at index: Int = 0) {
-            feedRequests[index](.success(feed))
-        }
-
-        func completeFeedLoadingWithError(at index: Int = 0) {
-            let error = NSError(domain: "test", code: 0)
-            feedRequests[index](.failure(error))
-        }
-
-        // MARK: - FeedImageDataLoader
-
-        private(set) var cancelledImageURLs = [URL]()
-        private var imageRequests = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)]()
-
-        private struct TaskSpy: FeedImageDataLoaderTask {
-            let cancelCallBack: () -> Void
-
-            func cancel() {
-                cancelCallBack()
-            }
-        }
-
-        var loadedImageURLs: [URL] {
-            imageRequests.map { $0.url }
-        }
-
-        func loadImageData(from url: URL, completion: @escaping (Result<Data, Error>) -> Void) -> FeedImageDataLoaderTask {
-            imageRequests.append((url, completion))
-            return TaskSpy { [weak self] in self?.cancelledImageURLs.append(url) }
-        }
-
-        func completeImageLoading(with imageData: Data = Data(), at index: Int = 0) {
-            imageRequests[index].completion(.success(imageData))
-        }
-
-        func completeImageLoadingWithError(at index: Int = 0) {
-            let error = NSError(domain: "test", code: 0)
-            imageRequests[index].completion(.failure(error))
-        }
-    }
-}
-
-// MARK: - FeedViewController helpers
-
-private extension FeedViewController {
-    func simulateUserInitiatedFeedReload() {
-        refreshControl?.simulatePullToRefresh()
-    }
-
-    var isShowingLoadingIndicator: Bool {
-        refreshControl?.isRefreshing ?? false
-    }
-
-    func numberOfRenderedFeedImageViews() -> Int {
-        tableView(self.tableView, numberOfRowsInSection: feedImagesSection)
-    }
-
-    var feedImagesSection: Int { 0  }
-
-    func feedImageView(at index: Int) -> UITableViewCell? {
-        let dataSource = tableView.dataSource
-        let indexPath = IndexPath(row: index, section: 0)
-        return dataSource?.tableView(tableView, cellForRowAt: indexPath)
-    }
-
-    @discardableResult
-    func simulateImageFeedViewVisible(at index: Int) -> FeedImageCell? {
-        feedImageView(at: index) as? FeedImageCell
-    }
-
-    func simulateImageFeedViewIsHidden(at index: Int) {
-        let view = feedImageView(at: index)
-
-        let delegate = tableView.delegate
-        let indexPath = IndexPath(row: index, section: feedImagesSection)
-        delegate?.tableView?(tableView, didEndDisplaying: view!, forRowAt: indexPath)
-    }
-
-    func simulateFeedImageViewNearVisible(at row: Int) {
-        let datasource = tableView.prefetchDataSource
-        let index = IndexPath(row: row, section: feedImagesSection)
-        datasource?.tableView(tableView, prefetchRowsAt: [index])
-    }
-
-    func simulateFeedImageViewNotNearVisible(at row: Int) {
-        simulateImageFeedViewVisible(at: row)
-
-        let datasource = tableView.prefetchDataSource
-        let index = IndexPath(row: row, section: feedImagesSection)
-        datasource?.tableView?(tableView, cancelPrefetchingForRowsAt: [index])
-    }
-}
-
-// MARK: - Cell Helpers
-
-private extension FeedImageCell {
-    var isShowingLocation: Bool {
-        !locationContainer.isHidden
-    }
-
-    var descriptionText: String? {
-        descriptionLabel.text
-    }
-
-    var locationText: String? {
-        locationLabel.text
-    }
-
-    var isShowingImageLoadingIndicator: Bool {
-        feedImageContainer.isShimmering
-    }
-
-    var renderedImage: Data? {
-        feedImageView.image?.pngData()
-    }
-
-    var isShowingRetryAction: Bool {
-        !feedImageRetryButton.isHidden
-    }
-
-    func simulateRetryAction() {
-        feedImageRetryButton.simulateTap()
-    }
-}
-
-// MARK: - UIKit Helpers
-
-extension UIRefreshControl {
-    func simulatePullToRefresh() {
-        allTargets.forEach { target in
-            actions(forTarget: target, forControlEvent: .valueChanged)?.forEach {
-                (target as NSObject).perform(Selector($0))
-            }
-        }
-    }
-}
-
-private extension UIButton {
-    func simulateTap() {
-        allTargets.forEach { target in
-            actions(forTarget: target, forControlEvent: .touchUpInside)?.forEach {
-                (target as NSObject).perform(Selector($0))
-            }
-        }
-    }
-}
-
-extension UIImage {
-    static func make(withColor color: UIColor) -> UIImage {
-        let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
-        UIGraphicsBeginImageContext(rect.size)
-        let context = UIGraphicsGetCurrentContext()!
-        context.setFillColor(color.cgColor)
-        context.fill(rect)
-        let img = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        return img!
-    }
 }
