+++ EssentialFeed/EssentialFeediOSTests/Helpers/FeedImageCell+TestHelpers.swift
@@ -1,0 +1,32 @@
+import Foundation
+import EssentialFeediOS
+
+extension FeedImageCell {
+    var isShowingLocation: Bool {
+        !locationContainer.isHidden
+    }
+
+    var descriptionText: String? {
+        descriptionLabel.text
+    }
+
+    var locationText: String? {
+        locationLabel.text
+    }
+
+    var isShowingImageLoadingIndicator: Bool {
+        feedImageContainer.isShimmering
+    }
+
+    var renderedImage: Data? {
+        feedImageView.image?.pngData()
+    }
+
+    var isShowingRetryAction: Bool {
+        !feedImageRetryButton.isHidden
+    }
+
+    func simulateRetryAction() {
+        feedImageRetryButton.simulateTap()
+    }
+}
