+++ EssentialFeed/EssentialFeediOSTests/Helpers/FeedViewController+LoaderSpy.swift
@@ -1,0 +1,59 @@
+import EssentialFeediOS
+import EssentialFeed
+
+extension FeedViewControllerTests {
+    final class LoaderSpy: FeedLoader, FeedImageDataLoader {
+        
+        // MARK: - FeedLoader
+        
+        var loadFeedCallCount: Int {
+            feedRequests.count
+        }
+        
+        private var feedRequests = [(FeedLoader.Result) -> Void]()
+        
+        func load(completion: @escaping (FeedLoader.Result) -> Void) {
+            feedRequests.append(completion)
+        }
+        
+        func completeFeedLoading(with feed: [FeedImage] = [], at index: Int = 0) {
+            feedRequests[index](.success(feed))
+        }
+        
+        func completeFeedLoadingWithError(at index: Int = 0) {
+            let error = NSError(domain: "test", code: 0)
+            feedRequests[index](.failure(error))
+        }
+        
+        // MARK: - FeedImageDataLoader
+        
+        private(set) var cancelledImageURLs = [URL]()
+        private var imageRequests = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)]()
+        
+        private struct TaskSpy: FeedImageDataLoaderTask {
+            let cancelCallBack: () -> Void
+            
+            func cancel() {
+                cancelCallBack()
+            }
+        }
+        
+        var loadedImageURLs: [URL] {
+            imageRequests.map { $0.url }
+        }
+        
+        func loadImageData(from url: URL, completion: @escaping (Result<Data, Error>) -> Void) -> FeedImageDataLoaderTask {
+            imageRequests.append((url, completion))
+            return TaskSpy { [weak self] in self?.cancelledImageURLs.append(url) }
+        }
+        
+        func completeImageLoading(with imageData: Data = Data(), at index: Int = 0) {
+            imageRequests[index].completion(.success(imageData))
+        }
+        
+        func completeImageLoadingWithError(at index: Int = 0) {
+            let error = NSError(domain: "test", code: 0)
+            imageRequests[index].completion(.failure(error))
+        }
+    }
+}
